requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pypi.org==0.0.0
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==0.16.1
zipp==3.1.0
